Add an animation example
A frequently asked question is: how do I animate drawing on a GtkGLArea
widget, if I don't have a draw loop like SDL does?

The answer is: GTK already has a draw loop, so all you need to do is
hook into it. We use the tick callback and the frame clock in order to
get timings driven (in the best case scenario) by the compositor — i.e.
by the display's vertical refresh rate.